backend/requirements.txt
Django>=5.1.0,<5.2
djangorestframework>=3.15.0
django-filter>=24.0
django-cors-headers>=4.3.0
cryptography>=42.0.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0
psycopg2-binary>=2.9.9
redis>=5.0.0
celery>=5.3.0
simpleeval>=0.9.13
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
drf-nested-routers>=0.94.1
dj-database-url>=2.1.0
httpx
django-celery-beat

# AI Agents - LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-qdrant>=0.2.0

# Vector Store
qdrant-client>=1.11.0

# Nested Routers for DRF
drf-nested-routers>=0.94.0

# HTTP Client
httpx>=0.27.0

# Jinja2 for templates
Jinja2>=3.1.0

# Pydantic for schemas
pydantic>=2.0.0
aiosmtplib>=3.0.0
